Move templated methods to `hpp` file.
