+++ include/Thread/ThreadPool.hpp
@@ -36,6 +36,28 @@ namespace LilWrapper
       template <typename O>
         void schedule(void (O::*)(), O *obj);
   };
+
+  template <typename F>
+    void ThreadPool::schedule(F function)
+    {
+      ThreadStarter *starter = new ThreadFunctor<F>(function);
+      this->_tasks.push(starter);
+    }
+
+  template <typename F, typename A>
+    void ThreadPool::schedule(F func, A arg)
+    {
+      ThreadStarter *starter = new ThreadFunctorWithArg<F, A>(func, arg);
+      this->_tasks.push(starter);
+    }
+
+  template <typename O>
+    void ThreadPool::schedule(void (O::*func)(), O *object)
+    {
+      ThreadStarter *starter = new ThreadMemberFunction<O>(func, object);
+      this->_tasks.push(starter);
+    }
+
 }
 
 #endif /* !THREADPOOL_HH_ */
